+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v4.0.0
+## May 29, 2024
+- Updated dependencies to address vulnerabilities in npm packages.
+- Removed unit testing capabilities to remove additional vulnerabilities.
+- Added React as a development option.
+- Added support for advanced group filter testing in local environment.
+
 # v3.4.3
 ## March 23, 2023
 - Added build-dev script for building and packaging the add-in for testing when uploading the files directly to MyG.